requirements.txt
alabaster==0.7.12
amqp==5.0.9
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.4.1
astroid==2.9.2
attrs==21.4.0
Babel==2.9.1
backcall==0.2.0
billiard==3.6.4.0
black==21.12b0
celery==5.2.3
certifi==2021.10.8
cffi==1.15.0
cfgv==3.3.1
charset-normalizer==2.0.10
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
coverage==6.2
crispy-bootstrap5==0.6
cryptography==36.0.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
distlib==0.3.4
Django==3.2.10
django-allauth==0.47.0
django-celery-beat==2.2.1
django-cors-headers==3.10.1
django-coverage-plugin==2.0.2
django-crispy-forms==1.13.0
django-debug-toolbar==3.2.4
django-environ==0.8.1
django-extensions==3.1.5
django-model-utils==4.2.0
django-oauth-toolkit==1.6.1
django-redis==5.2.0
django-stubs==1.9.0
django-stubs-ext==0.3.1
django-timezone-field==4.2.1
djangorestframework==3.13.1
djangorestframework-stubs==1.4.0
docutils==0.17.1
factory-boy==3.2.1
Faker==11.3.0
filelock==3.4.2
flake8==4.0.1
flake8-isort==4.1.1
flower==1.0.0
h11==0.12.0
hiredis==2.0.0
httptools==0.3.0
humanize==3.13.1
identify==2.4.2
idna==3.3
imagesize==1.3.0
iniconfig==1.1.1
ipdb==0.13.9
ipython==7.31.0
isort==5.10.1
itypes==1.2.0
jedi==0.18.1
Jinja2==3.0.3
jwcrypto==1.0
kombu==5.2.3
lazy-object-proxy==1.7.1
livereload==2.6.3
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mccabe==0.6.1
mypy==0.930
mypy-extensions==0.4.3
nodeenv==1.6.0
oauthlib==3.1.1
packaging==21.3
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
platformdirs==2.4.1
pluggy==1.0.0
pre-commit==2.16.0
prometheus-client==0.12.0
prompt-toolkit==3.0.24
psycopg2==2.9.3
ptyprocess==0.7.0
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
Pygments==2.11.2
PyJWT==2.3.0
pylint==2.12.2
pylint-celery==0.3
pylint-django==2.4.4
pylint-plugin-utils==0.7
pyparsing==3.0.6
pytest==6.2.5
pytest-django==4.5.2
pytest-sugar==0.9.4
python-crontab==2.6.0
python-dateutil==2.8.2
python-dotenv==0.19.2
python-slugify==5.0.2
python3-openid==3.2.0
pytz==2021.3
PyYAML==6.0
redis==4.1.0
requests==2.27.1
requests-oauthlib==1.3.0
six==1.16.0
snowballstemmer==2.2.0
Sphinx==4.3.2
sphinx-autobuild==2021.3.14
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.2
termcolor==1.1.0
testfixtures==6.18.3
text-unidecode==1.3
toml==0.10.2
tomli==1.2.3
tornado==6.1
traitlets==5.1.1
types-pytz==2021.3.3
types-PyYAML==6.0.1
typing-extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.7
uvicorn==0.16.0
uvloop==0.16.0
vine==5.0.0
virtualenv==20.13.0
watchgod==0.7
wcwidth==0.2.5
websockets==10.1
Werkzeug==2.0.2
wrapt==1.13.3
